apps/fhe-service/requirements.txt
# FHE Service Dependencies
# Concrete Python for FHE operations
concrete-python>=2.0.0

# FastAPI for REST API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Pydantic for data validation
pydantic>=2.5.0

# Environment variables
python-dotenv>=1.0.0

# CORS support
python-multipart>=0.0.6

# Logging
loguru>=0.7.2

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
requests>=2.31.0

